fix public key file parsing
The comment field in a public key file is optional.
If you e. g. use OpenSSH to extract a public key file from a private key file, the field is not set.
Previously the public key file parsing was not able to handle this properly, but failed with an exception.