prefer --admin if also --write is given
When using git bb add-privilege and git bb invite and --admin as well as --write was specified, the user only got write permissions.
Now --admin takes precedence, so the user is getting all privileges that were specified on the commandline, as the commandline parser does not varify that only one of the privileges is given on the commandline.