add possibilities to specify messages on the commandline
For all commands that requested a message from the user
(issue, pull-request, merge-request and release)
there are now five additional options.
"--message" to specify the message on the commandline.
"--file" to use the content of a file or standard input as message.
"--template" to use the content of a file as a start for the message.
"--reuse-message" to use the message of a commit-ish for the message.
"--edit" to further edit "--message", "--file" or "--reuse-message" messages.