fix the issues command with parent flag
When calling the issues command for any spindle with the --parent flag and for a specific repository (opposed to all user repositories), the --parent flag was evaluated twice. Once inside the self.repository method and ones in the following for loop. This resulted in the grand-parent issues being shown if a grand-parent exists instead of the parent issues.
Now the parent issues are displayed correctly.