enhance forks listing for GitHub and Bitbucket
Up to now only the direct forks of the parent repository were listable (or the own if no parent existed).
Now by default the direct forks of the current repository are listed and there are options to either list the parent forks or the root repositories forks and to recursive list all repos down the fork tree. With --root and --recursive used together you get all repositories in the network of the repository in question.