search in the whole network when adding a remote for GitHub or Bitbucket
Up to now only the direct forks of the parent repository were searched for the remote to be added (or the own if no parent existed).
Now by default the direct forks of the current repository and their descendents are searched first and then the parent and its other forks and so on are searched until the fork is found or the network was searched completely but nothing found.