fix setup-goblet for GitHub and provide it for the others
setup-goblet for GitHub set the wrong config keys and the description and URLs were taken from the grand-parent if available else from the parent if available else from the current repository.
Now the correct config keys are set, the description is used from the root repository of the network, the URLs from the current repository and GitLab and Bitbucket also have the set-goblet command and not only the --goblet parameter to mirror.