fetch pull requests into the remotes namespace for all added remotes
Up to now for GitHub the pull requests were fetched into refs/pull/*/head (also for BitBucket when applying a PR), for GitLab the merge requests were fetched into refs/merge-requests/*/head.
This has some drawbacks, e. g. if you use "set-origin" or "fork", the pull requests are still the one from the previous remote, if you rename or delete a remote, the pull request refs are not updated accordingly, ...

Now pull and merge requests for all remotes that are added or set by add-remote, set-origin, clone, create or fork are now fetched also and stored under "refs/remotes/<remote>/(pull|merge)-requests/<number>".
This way you can have PRs and MRs for upstream, origin and also forks if someone made a PR / MR to one fork in the network and you want to merge it into your repo.
You can now also use bash completion for "git checkout" and similar, with those refs more easily as they are provided immediately in the completion.
The "head" in the end of the refs is not really helpful, as only the heads are fetched anyway, so they are left out now.
If you rename a remote or delete a remote or fork the repo and thus rename origin to upstream, the PRs / MRs will now also move along with the deletion or renaming.