add option and parameter to not open the browser
"<spindle> browse" was not too useful when working with Git on some headless remote server.
Now there is a parameter "--no-browser" and a boolean configuration option "gitspindle.no-browser" that can be set to prevent the browser from being launched. Instead a message is printed that contains the URL and you can open this in your local browser. For consistency reason the "hub render" command also honors this option and got the parameter, though you could already achieve something similar with "--save". The actual difference is, that with "--no-browser" the file is automatically created in the temporary directory and the URL to the file is printed, while "--save" does not output anything.