when creating a Bitbucket repo for a team also specify project
In Bitbucket team repositories are always organized in projects.
If you try to create a repository without project it lands automatically in the oldest available project.
If there is no project yet, it lands in a newly created unnamed project.
Now when creating a team repository you also have to specify the project under which the repository should be stored.