only ask the two-factor authentication code interactively
When executing the two-factor tests, the two relevant tests that need a token asked for all information (username, password, token), as the communication was simply left to the production code.
Now the username and password are taken from the .gitspindle file and only the token is asked for by the test itself and then given to the production code like in all other places in the tests.