do not require tested servers to be present in known_hosts file
If you do not have all SSH keys of the tested servers in your known_hosts file, you are constantly asked for confirmation during the test executions.
Now the StrictHostKeyChecking option of SSH is disabled which makes the tests to work without always asking even if the SSH keys are not present in the known_hosts file.