do not use spindle_host to determine spindle_remote
Up to now spindle_remote uses spindle_host to determine the result. This does not work nicely for the local GitLab instance as it needs to be configured including protocol and even if protocol would be stripped requires the local GitLab instance to be reachable via an URL that starts with "gitlab.".
Now an own case construct is used in spindle_remote to make it independent from the hostnames.