make tests independent of fortune
The tests used the fortune executable to produce some content to be committed. This is unfortunate, as the fortune tool is not necessarily available everywhere and is nowhere listed as dependency.
As it is not really necessary to use the fortune tool, now just "echo", "date" and the "commit_count" variable are used to produce content to be committed.