make tests independent from local Git configuration
The test commit helper function created a commit. Git expects "user.name" and "user.email" to be set. If it does not find them, it tries to calculate something which does not always work for user.email, e. g. if you have a local system without domain name, then it would generate something like "vampire@vm-vampire-ubuntu." which is not a valid e-mail address and thus Git refuses to continue.
Now there is a hardcoded user.name and user.email set for the commit command that makes the tests independent of local Git and host configuration.