improve webbrowser fooling script
The xdg-open script that was intended to fool the Python webbrowser module echoed the input and exited immediately.
Unfortunately the webbrowser module checks whether the invoked process is still alive after invoking it.
As the process already finished when this check is done, the next browser command is tried and the URL is opened in a real webbrowser.
Now the script waits for 10 seconds to successfully fool the webbrowser module.