fix README for tests
The README for tests had an unnecessary mention of the reuirement that the account names must contain "test", as there is also the requirement that all the accounts start with "git-spindle-test-", otherwise the test-cleanup command will choke and fail.
Also the README missed information about the mandatory org / public group / team that needs to be present and configured in .gitspindle.
Also TWOFACTORTESTS and TEST_OPTS were not documented in the README.
Now all these are fixed.