wait for creation or forking to complete for pushing
If you try to push to a repository, while the repository is still being created or forked on GitLab due to its asynchronous working mode, the operation fails.
It is now tried for 120 times, waiting a second between the tries to repeat the operation if it fails.
Since a recent update to GitLab, the tests almost never run through due to this behavior. With this change they are working again.